pyproject.toml
[tool.poetry]
name = "llm-test-automation"
version = "0.1.0"
description = "NC_llm_judge_testcase.xlsx 기반 LLM 테스트 자동화 도구"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"  # Python 3.11 버전 지정
streamlit = "^1.31.0"
pandas = "^2.1.4"
numpy = "^1.24.3"  # 컴파일 없이 설치 가능한 안정 버전
plotly = "^5.18.0"
pyyaml = "^6.0.1"
openpyxl = "^3.1.2"
python-dateutil = "^2.8.2"
langchain = "^0.1.9"
langchain-openai = "^0.0.7"
langgraph = "^0.0.23"
openai = "^1.13.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"